refactor:(shared): utils tests
